microservices/bicycle/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
alembic==1.13.0
psycopg[binary]==3.1.18
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.0
pika==1.3.2
requests==2.31.0
PyJWT==2.8.0
psycopg2-binary
